pyproject.toml
[project]
name = "nonebot_plugin_picsbank"
version = "0.1.4"
description = ""
authors = [
    {name = "rf_tar_railt", email = "3165388245@qq.com"},
    {name = "Alex Newton", email = "sharenfan222@gmail.com"},
]
dependencies = [
    "ujson<5.0.0,>=4.0.2",
    "nonebot2[aiohttp]>=2.3.0",
    "numpy<2.0.0,>=1.21.4",
    "Pillow<10.0.0,>=8.2.0",
    "opencv-python>=4.5.4",
    "nonebot-plugin-localstore>=0.6.0",
    "nonebot-plugin-waiter>=0.4.1",
    "nonebot-plugin-alconna>=0.45.4",
    "nonebot-plugin-userinfo>=0.2.4",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}
keywords = ["nonebot", "nonebot2", "picsbank"]

[project.urls]
homepage = "https://github.com/MaxCrazy1101/nonebot_plugin_picsbank"
repository = "https://github.com/MaxCrazy1101/nonebot_plugin_picsbank"



[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "nonebot-adapter-onebot>=2.4.3",
    "nonebot2[fastapi,uvicorn]>=2.3.0",
    "black>=24.4.2",
    "isort>=5.13.2",
    "ruff>=0.4.4",
]

[tool.pdm.build]
includes = ["src"]
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.black]
line-length = 120
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 120
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["C901", "T201", "E731", "E402"]

[tool.pyright]
pythonVersion = "3.9"
pythonPlatform = "All"
defineConstant = { PYDANTIC_V2 = true }
typeCheckingMode = "basic"
reportShadowedImports = false
disableBytesTypePromotions = true

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugin_dirs = ["src"]
builtin_plugins = []